Add the prototype version of Test::PerlRun
This is Dave's original, but with _run() rewritten to use avoid IPC::Cmd 'run',
which can't capture output on Win32 with core-only modules. Instead, _run() now
uses open on platforms that support '-|' (seemingly everything except Win32),
falling back to ``, using a temporary file in all cases if STDERR needs
capturing. This avoids complicated select code, and most of the quoting issues
test.pl faces, due to test.pl using `` on VMS. However, we need to avoid
newlines and double quotes in code passed via -e, as double-quoting is the
mechanism for escaping for the Win32 shell.

Additionally as we have direct access to $?, we can eliminate the hacky code to
reprocess IPC::Cmd::run()'s munging of the exit status, and can correctly
distinguish between a process which exited cleanly, and one which had a fatal
signal, although, frustratingly, it seems that kill 'TERM', $$; doesn't, on
Win32. POSIX it ain't.