Add the ability to pass data on STDIN to the perl run by Test::PerlRun
When passing a scalar containing code (instead of a hash reference) change the
implementation to pass the code as a program on STDIN, instead of using -e on
the command line. This allows "complex" code to be easily passed, without
needing to resort to temporary files in the caller. ("Complex" includes
common constructions such as double quotes and newlines, hence why this is
useful.)